add the transform_to_0_n_bins function
This was a step that was performed in the original step0; I am not
sure it's still needed, but I keep it because it's adding one bin at
the end. With this change I can exactly reproduce the output of the
original step0 (still a small diff in the fake syst histos).

Note that the suffix '_rebin' is moved from pt_slice_histoname to
transform_to_0_n_bins.


git-svn-id: svn+ssh://svn.cern.ch/reps/atlasphys-hsg4/Physics/Higgs/HSG4/LFV/Leplep/FitModel@234637 883ba7d9-fdd0-4202-9341-49aa55999ad8